Allow tagger script to access file metadata
Fixes PICARD-259 (and associated sub-tickets)

This moves the taggerscript evaluation to happen when a file is matched
(i.e. when a file is added to a track), where it has access to the both
the file metadata *and* the track metadata.

In order to ensure that the titles in the user interface are correct,
this is done by saving an extra copy of the original musicbrainz
metadata in track.orig_metadata. The taggerscript is run as per normal
on track.metadata, but the file matching will use track.orig_metadata
and re-run the taggerscript on the merged data.